% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-gwqs.R
\name{sr_gwqs}
\alias{sr_gwqs}
\title{Fitting Weighted Quantile Sum regression models}
\usage{
sr_gwqs(
  data,
  outcome = NULL,
  exposure = NULL,
  covariates = NULL,
  q = 4,
  b = 10,
  validation = 0.3,
  family = NULL,
  seed = 2023,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{exposure}{exposure variable names.}

\item{covariates}{covariate variable names.}

\item{q}{An integer to specify how mixture variables will be ranked, e.g. in
quartiles (q = 4), deciles (q = 10), or percentiles (q = 100). If q = NULL
then the values of the mixture variables are taken (these must be standardized).}

\item{b}{number of bootstrap samples used in parameter estimation.}

\item{validation}{percentage of the dataset to be used to validate the model.
If validation = 0 then the test dataset is used as validation dataset too.}

\item{family}{a character value that allows to decide for the glm: gaussian
for linear regression, binomial for logistic regression "multinomial" for
multinomial regression, poisson for Poisson regression, quasipoisson for
quasi-Poisson regression, "negbin" for negative binomial regression.}

\item{seed}{an integer value to fix the seed, if it is equal to NULL no seed
is chosen.}

\item{...}{additional arguments to be passed to the \code{\link[gWQS:main_functions]{gWQS::gwqs()}} function.}
}
\value{
a list.
}
\description{
Fitting Weighted Quantile Sum regression models
}
