% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_gwqs_barplot.R
\name{sr_gwqs_barplot}
\alias{sr_gwqs_barplot}
\title{Barplot for weights}
\usage{
sr_gwqs_barplot(
  object,
  label = FALSE,
  label.size = 12,
  label.digits = 3,
  bar.color = NULL,
  bar.width = 0.65,
  ...
)
}
\arguments{
\item{object}{an object from \link{sr_gwqs} function.}

\item{label}{logical, whether to show weight values.}

\item{label.size}{label szie.}

\item{label.digits}{label digits.}

\item{bar.color}{bar color.}

\item{bar.width}{bar width.}

\item{...}{unused.}
}
\value{
a ggplot2 object.
}
\description{
Barplot for weights
}
