% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_gwqs_coefs.R
\name{sr_gwqs_coefs}
\alias{sr_gwqs_coefs}
\title{Coefficients of gWQS regression}
\usage{
sr_gwqs_coefs(object, ...)
}
\arguments{
\item{object}{an object from \link{sr_gwqs} function.}

\item{...}{additional arguments to be passed to the \link[srmisc:typeset]{srmisc::typeset} function.}
}
\value{
a data frame.
}
\description{
Coefficients of gWQS regression
}
